+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/task/InferenceTask.kt
@@ -1,7 +1,7 @@
 package edu.wpi.axon.core.dsl.task
 
 import edu.wpi.axon.core.dsl.Import
-import edu.wpi.axon.core.dsl.variable.Code
+import edu.wpi.axon.core.dsl.Code
 import edu.wpi.axon.core.dsl.variable.InferenceSession
 import edu.wpi.axon.core.dsl.variable.ModelInputData
 import edu.wpi.axon.core.dsl.variable.Variable
