+++ core/core/src/main/kotlin/edu/wpi/axon/core/dsl/variable/ImageInputData.kt
@@ -1,5 +1,6 @@
 package edu.wpi.axon.core.dsl.variable
 
+import edu.wpi.axon.core.dsl.Code
 import edu.wpi.axon.core.dsl.Import
 
 /**
