+++ core/core/src/test/kotlin/edu/wpi/axon/core/dsl/CodeGraphTest.kt
@@ -4,7 +4,6 @@ import com.google.common.graph.EndpointPair
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import edu.wpi.axon.core.dsl.container.PolymorphicNamedDomainObjectContainer
-import edu.wpi.axon.core.dsl.variable.Code
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
