+++ core/core/src/test/kotlin/edu/wpi/axon/core/dsl/MockTask.kt
@@ -1,7 +1,6 @@
 package edu.wpi.axon.core.dsl
 
 import edu.wpi.axon.core.dsl.task.Task
-import edu.wpi.axon.core.dsl.variable.Code
 import edu.wpi.axon.core.dsl.variable.Variable
 import java.util.concurrent.CountDownLatch
